% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aggregate_eval.R
\name{aggregate_eval}
\alias{aggregate_eval}
\title{Aggregate multiple course evalautions}
\usage{
aggregate_eval(dat, id, lang = "de", x_labels = NULL)
}
\arguments{
\item{dat}{Data frame as returned from \code{\link{read_eval}}}

\item{id}{Optional numeric vector identifying a subset of data sets, which should be used.}

\item{lang}{Character string specifying the language of the variable names. Currently implemented only \code{"de"} for German.}

\item{x_labels}{Optional character vector supplying the names of the courses.
Number of elements must match the number of files.}
}
\value{
Returns a list of five elements:
\itemize{
 \item{\code{mean}: For each scale and each course evaluation, the mean across all participants}
 \item{\code{sd}: For each scale and each course evaluation, the SD across all participants}
 \item{\code{se}: For each scale and each course evaluation, the SE across all participants}
 \item{\code{N}: For each course evaluation, the number of participants}
 \item{\code{varnames}: For each scale, its label}
}
}
\description{
This function takes as input the output from \code{\link{read_eval}} and returns a list of aggregated data (e.g., means, standard errors).
}
\examples{
\dontrun{
dat_1 <- read_eval("./data/")      # read all files
res.1 <- aggregate_eval(dat_1)     # aggregate results for plotting
}
}

