% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/comb_eval.R
\name{comb_eval}
\alias{comb_eval}
\title{Combine and aggregate multiple course evalautions}
\usage{
comb_eval(dat, id)
}
\arguments{
\item{dat}{List of data frames as returned from \code{read_eval}}

\item{id}{Optional numeric vector identifying a subset of files, which should be read.}
}
\value{
Returns a list of five elements:
\describe{
 \item{mean}{For each scale and each course evaluation, the mean across all participants}
 \item{sd}{For each scale and each course evaluation, the SD across all participants}
 \item{se}{For each scale and each course evaluation, the SE across all participants}
 \item{N}{For each course evaluation, the number of participants}
 \item{labels}{For each scale, its label}
}
}
\description{
This function takes as input the output from \code{read_eval} and returns a list of aggregated data (e.g., means, standard errors).
}
\examples{
\dontrun{
dat.1 <- read_eval("./data/")            # read all files
res.1 <- comb_eval(dat.1)                # combine results for plotting
}
}

