% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_eval.R
\name{datatable_eval}
\alias{datatable_eval}
\title{HTML table widget of all courses using the DataTables library}
\usage{
datatable_eval(data, digits = 2)
}
\arguments{
\item{data}{Data frame as returned from \code{\link{join_eval}}}

\item{digits}{Numeric indicating the number of decimal places to be used. Passed to \code{\link[base]{round}}.}
}
\description{
This function takes a data frame from \code{\link{join_eval}} containing
data from multiple courses and displays summary statistics (e.g., mean) using
the DataTables library.
}
\examples{
\dontrun{
tmp1 <- read_eval("./data")
dat1 <- join_eval(tmp1)
datatable_eval(dat1)
}

dat1 <- sim_eval()
datatable_eval(dat1)
}
\seealso{
\code{\link[DT]{datatable}}.
}
