% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instevalR.R
\docType{package}
\name{instevalR}
\alias{instevalR}
\alias{instevalR-package}
\title{instevalR: An R Package For Summarizing And Plotting Several InstEvaL Course Evaluations}
\description{
InstEvaL is an online system for course evaluations, which is used at some
German universities (e.g., U Mannheim). The system
(\url{https://insteval.uni-mannheim.de/}) provides instructors with feedback for
every course that is evaluated and with the accompanying raw data. The
package instevalR uses the raw data of several course evaluations, combines
them, and provides a handful of interesting plots.
}
\section{Workflow}{

\enumerate{
 \item{You have \code{R} and \code{instevalR} installed. Furthermore, please download the files of your course evaluations (log in to InstEvaL -> Results -> Raw data) and put them into a single directory.}
 \item{Then you need to import all your data into R using the function \code{\link{read_eval}}.}
 \item{Afterwards, you need to merge all course evaluations using \code{\link{join_eval}}.}
 \item{Finally, you can plot your data using \code{\link{plot_eval}}. Play around with the argument \code{plottype} and generate different plots.}
}
}

\examples{
## EITHER ## read own data
\dontrun{
tmp1 <- read_eval()
dat1 <- join_eval(tmp1)
}

## OR ## simulate data
dat1 <- sim_eval()

# plot
plot_eval(dat1)
}
