% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_eval.R
\name{join_eval}
\alias{join_eval}
\title{Join multiple course evaluations into one data frame}
\usage{
join_eval(data, subset = NULL, lang = c("de", "en"), names = NULL,
  wide_format = FALSE)
}
\arguments{
\item{data}{Data frame as returned from \code{\link{read_eval}}}

\item{subset}{Optional numeric vector identifying a subset of data sets,
which should be used (e.g., \code{c(1, 3:5)}).}

\item{lang}{Character string specifying the language of the variable names.}

\item{names}{Optional character vector supplying the names of the courses.
Number of elements must match the number of files.}

\item{wide_format}{Logical indicating whether data should be returned in wide format.}
}
\value{
A data frame (more precisely, a tibble) with scores for all scales and persons.
}
\description{
This function takes a list of data frames from \code{\link{read_eval}} and
returns a single data frame with appropriate labels.
}
\examples{
\dontrun{
tmp1 <- read_eval("./data/")      # read all files
dat1 <- join_eval(tmp1)          # combine data frames for plotting
}
}
