% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eval.R
\name{plot_eval}
\alias{plot_eval}
\title{Plot multiple course evalautions}
\usage{
plot_eval(data = NULL, plottype = c("barplot", "points", "boxplot",
  "violin"), domains = 2, topics = NULL, topic_as = c("facet", "group",
  "x-axis"), course_as = c("x-axis", "facet", "group"), add_line = FALSE,
  errorbar = TRUE, CI = 0.95, grey = FALSE, lang = NULL,
  xvar_as_legend = FALSE, control = list(), ...)
}
\arguments{
\item{data}{Data frame as returned from \code{\link{join_eval}} (or from
\code{\link{sim_eval}}).}

\item{plottype}{Character, either \code{"barplot"}, \code{"points"}, \code{"boxplot"}, or \code{"violin"}.}

\item{domains}{Numeric of length >= 1, specifying the domains to plot.
\itemize{
 \item{1: Overall (one scale, namely, average of \code{domains = 2})}
 \item{2: Overall, (i.e., didactic skills, grade lecturer, grade course, and comparison with other courses)}
 \item{3: Different scales (e.g., structure, sympathy, motivation)}
 \item{4: Settings (e.g., temperature, acoustics, media use)}
 \item{5: Students' presentations}
 \item{6: Misc}
}}

\item{topics}{Character vector of length >= 1, optional. Used to select a
subset of topics present in \code{data$topic}, e.g., \code{topics =
c("Overall", "Motivation")}. Overrides argument \code{domains}.}

\item{topic_as}{Character. Should each topic be plotted using a different
\code{"facet"} (panel), different \code{"group"}, or a different value on
the \code{"x-axis"}?}

\item{course_as}{Character. Should each course/file be plotted using a
different value on the \code{"x-axis"}, a different \code{"facet"} (panel),
or a different \code{"group"}?}

\item{add_line}{Logical. Should a connecting line be added to the plot?}

\item{errorbar}{Logical indicating whether error bars representing a
confidence interval should be plotted or not.}

\item{CI}{Numeric. Confidence level of error bars, often \code{.95} for a 95\% CI.}

\item{grey}{Logical indicating whether colors should be replaces with shades
of grey.}

\item{lang}{Character string specifying the language of the variable names.}

\item{xvar_as_legend}{Logical indicating whether a legend should be plotted
for the variable on the x-axis.}

\item{control}{Optional list to fine-tune the plot.
\describe{
 \item{nrow}{Numeric. Number of rows of a facetted plot. Passed to \code{\link[ggplot2]{facet_wrap}}.}
 \item{ncol}{Numeric. Number of columns of a facetted plot. Passed to \code{\link[ggplot2]{facet_wrap}}.}
 \item{title}{Character. Passed to \code{\link[ggplot2]{labs}}.}
 \item{legend_title}{Character. Passed to \code{\link[ggplot2]{labs}}.}
 \item{x_lab}{Character. Passed to \code{\link[ggplot2]{labs}}.}
 \item{y_lab}{Character. Passed to \code{\link[ggplot2]{labs}}.}
 \item{angle, hjust}{Numeric. Adjustment of the labels of the ticks on the x-axis. Passed to \code{\link[ggplot2]{element_text}}.}
 \item{errorbar_width}{Numeric. Width of the whiskers. Passed to \code{\link[ggplot2]{geom_errorbar}}.}
 \item{ylim}{Numeric of length <= 2. Limits of the y-axis. Passed to \code{\link[ggplot2]{coord_cartesian}}.}
 \item{alpha}{Numeric. Alpha transparency.}
 \item{size}{Numeric. Factor modifying the size of points and errorbars, defaulting to 1.}
 \item{shape}{Numeric/Character. Shape of points. Passed to \code{\link[ggplot2]{geom_point}}.}
 \item{dodge_width}{Numeric. How far away from each other should different groups be plotted? Passed to \code{\link[ggplot2]{position_dodge}}.}
}}

\item{...}{Other parameters passed to---depending on \code{plottype}---\code{\link[ggplot2]{geom_col}}, \code{\link[ggplot2]{geom_point}}, \code{\link[ggplot2]{geom_boxplot}}, or \code{\link[ggplot2]{geom_violin}}.}
}
\description{
This function takes as input a data frame from \code{\link{join_eval}} (or
from \code{\link{sim_eval}}) and plots the results. Choose the type of plot
using the argument \code{plottype}. Furhtermore,  decide how the different
courses and the different topics should be represented in the plot using
\code{course_as} and \code{topic_as}, namely, either on the x-axis, in
different facets/panels, or using different colors/groups.
}
\examples{
\dontrun{
tmp1 <- read_eval("./data")
dat1 <- join_eval(tmp1)
}

# Simulate data
dat1 <- sim_eval()

plot_eval(dat1)

plot_eval(dat1, plottype = "points", domains = 1)

plot_eval(dat1, errorbar = FALSE)

plot_eval(dat1, topics = c("Motivierung", "Note Dozent"), plottype = "boxplot")

plot_eval(dat1, plottype = "points", topic_as = "x-axis", course_as = "group", add_line = TRUE)

plot_eval(dat1, domains = 1:2, plottype = "violin", xvar_as_legend = TRUE,
          control = list(title = "My Title")) +
    ggplot2::theme_dark()
}
