% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_eval.R
\name{read_eval}
\alias{read_eval}
\title{Read insteval data from directory}
\usage{
read_eval(directory, id)
}
\arguments{
\item{directory}{Character string of path name, where the *.csv-files are stored.}

\item{id}{Optional numeric vector identifying a subset of files, which should be read.}
}
\value{
Returns a list of data frames, one for each *.csv-file.
}
\description{
Reads all *.csv-files from a given directory and returns all content in a single list.
}
\examples{
\dontrun{
dat.1 <- read_eval("./data/")            # read all files
dat.1 <- read_eval("./data/", id = 1:5)  # read first 5 files (sorted by file name)
}
}

