% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_eval.R
\name{shiny_eval}
\alias{shiny_eval}
\title{Interactive shiny app to plot several course evaluations}
\usage{
shiny_eval(display.mode = "normal", ...)
}
\arguments{
\item{display.mode}{The mode in which to display the application. If set to
the value \code{"showcase"}, shows application code and metadata from a
\code{DESCRIPTION} file in the application directory alongside the
application. If set to \code{"normal"}, displays the application normally.
Defaults to \code{"auto"}, which displays the application in the mode given
in its \code{DESCRIPTION} file, if any.}

\item{...}{Parameters passed to \code{\link[shiny]{runApp}}}
}
\description{
Launches an interactive shiny app to upload the CSV-files of several course
evaluations and plot them. This can be used as an alternative to
\code{\link{plot_eval}}.
}
\examples{
\dontrun{
library(instevalR)
shiny_eval()
}
}
