% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_eval.R
\name{plot_eval}
\alias{plot_eval}
\title{Plot multiple course evalautions}
\usage{
plot_eval(x, plottype = 1, subscale = NULL, error_bars = TRUE,
  ci = 0.95, x.labels = NULL, pdf = FALSE, col = NULL,
  col.axis = "salmon", alpha = 0.25, lwd = 3, main = NULL, pch = 20,
  type = "b", ylim = NULL, plot.legend = TRUE, ...)
}
\arguments{
\item{x}{List of data frames as returned from \code{\link{aggregate_eval}}.}

\item{plottype}{Integer between \code{1} and \code{4}, selects the type of plot(s):
\describe{
 \item{\code{plottype = 1}}{1 plot of 4 variables, i.e., the first four scales of "Gesamtbewertung"}
 \item{\code{plottype = 2}}{1 plot of 1 variable, i.e., the scale "5: Gesamt"}
 \item{\code{plottype = 3}}{33 plots of all 33 variables arranged in 7 plot windows}
 \item{\code{plottype = 4}}{1 plot of a single, user-seleted variable (see \code{subscale})}
}}

\item{subscale}{Integer indicating the number of the scale to be plotted (if
\code{plottype = 4}).}

\item{error_bars}{Logical indicating whether error bars representing a
confidence interval should be plotted or not.}

\item{ci}{Numeric. Confidence level, often .95 for a 95\% CI.}

\item{x.labels}{Optional character vector with the labels of the tick marks
of the x-axis, typically names of courses/semester. If \code{NULL}, this is
borrowed from the names of the *.csv-files.}

\item{pdf}{Logical. If \code{TRUE}, the plots are written to a pdf-file.}

\item{col}{Character string. The color to be used for the lines and error bars, may be a vector of length 4 if \code{plottype = 1}.}

\item{col.axis}{Character. If \code{plottype = 3} and \code{ylim = NULL} (its
default), most of the plots have black y-axis annotations. But some of the
plots differ in their \code{ylim}-values and their y-axis annotations are
printed in a different color. Can be changed to \code{"black"} to override
this behavior.}

\item{alpha}{Numeric. Transparency value for the error bars (\code{0} means
fully transparent and \code{1} means opaque).}

\item{lwd}{The line width, a positive number.}

\item{main}{a main title for the plot, see also \code{\link{title}}.}

\item{pch}{Either an integer specifying a symbol or a single character to be
used as the default in plotting points. See \code{\link{points}} for possible values and
their interpretation.}

\item{type}{1-character string giving the type of plot desired.  The
    following values are possible, for details, see \code{\link{plot}}:
    \code{"p"} for points, \code{"l"} for lines,
    \code{"b"} for both points and lines,
    \code{"c"} for empty points joined by lines,
    \code{"o"} for overplotted points and lines,
    \code{"s"} and \code{"S"} for stair steps and
    \code{"h"} for histogram-like vertical lines.  Finally,
    \code{"n"} does not produce any points or lines.}

\item{ylim}{the y limits of the plot.}

\item{plot.legend}{Logical. May be set to \code{FALSE} in order to specify a user-defined legend afterwards.}

\item{...}{Other \link{graphical parameters} passed to \code{\link[graphics]{lines}} or \code{\link[graphics]{title}}.}
}
\description{
This function takes as input the output from \code{\link{aggregate_eval}} and plots the results.
}
\examples{
\dontrun{
dat.1 <- read_eval("./data/")      # read all files
res.1 <- aggregate_eval(dat.1)     # aggregate results for plotting

# Default: Plot of four main scales with 95\% CI
plot_eval(res.1)

# Plot only the scale "Gesamt", with 90\% CI
plot_eval(res.1, plottype = 2, ci = .9)

# Plot all scales and save the plots in a pdf file
plot_eval(res.1, plottype = 3, pdf = TRUE)

# Plot a specific scale (#2) with user-defined parameters
plot_eval(res.1, plottype = 4, subscale = 2, lwd = 2, lty = 2, col = "blue",
    ylim = c(1, 6), ylab = "Scale", x.labels = paste("Course", 1:5))
}
}

