% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_eval.R
\name{read_eval}
\alias{read_eval}
\title{Read insteval data from directory}
\usage{
read_eval(path = NULL, names = NULL, pattern = NULL, recursive = FALSE,
  shiny = FALSE)
}
\arguments{
\item{path}{If \code{NULL}, the user is prompted to choose the files
interactively; this is only available on Windows. Otherwise, a character
vector containing path name(s), pointing to the CSV-files. The files
probably have names ending with "evaluationen.csv". Make sure to put only
those files in the directory and no other files.}

\item{names}{Optional character vector supplying the names of the courses.
Number of elements must match the number of files.}

\item{pattern}{An optional regular expression passed to
\code{\link[base]{list.files}}. Can be used to read data only from files
with a matching filename.}

\item{recursive}{Logical passed to \code{\link[base]{list.files}}. Should the
file seach recurse into directories?}

\item{shiny}{Logical. For internal use only.}
}
\value{
List of data frames.
}
\description{
Reads all CSV-files from a given directory and returns all data in a single list
}
\examples{
\dontrun{
dat1 <- read_eval()

# Specify own names and read only files matching a specific pattern
dat1 <- read_eval("./data", names = c("Lab1", "Lab2"), pattern = "lab")
}
}
