% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_eval.R
\name{read_eval}
\alias{read_eval}
\title{Read insteval data from directory}
\usage{
read_eval(directory, id, x.labels = NULL)
}
\arguments{
\item{directory}{Character string of path name, where the *.csv-files are
stored. (The files probably have a name like
"InstEvaL-Rohdaten-vlg_XXXXX-evaluationen.csv", make sure to put only those
files in the directory and no other files.)}

\item{id}{Optional numeric vector identifying a subset of files, which should be used.}

\item{x.labels}{Optional character vector supplying the names of the courses.
Number of elements must match the number of files to read in.}
}
\value{
Returns the data from all *.csv-file in one object.
}
\description{
Reads all *.csv-files from a given directory and returns all content in a single object.
}
\examples{
\dontrun{
dat.1 <- read_eval("./data/")               # read all files
dat.1 <- read_eval("./data/", id = 1:5)     # read first 5 files
}
}

