% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_eval.R
\name{read_raw}
\alias{read_raw}
\title{Read from CSV or excel file}
\usage{
read_raw(file)
}
\arguments{
\item{file}{Character. The name of the file}
}
\value{
Data frame
}
\seealso{
\code{\link[utils]{read.table}}
 \code{\link[readxl]{read_excel}}
}
