% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sim_eval.R
\name{sim_eval}
\alias{sim_eval}
\title{Simulate data for mock course evaluations}
\usage{
sim_eval(N.files = 5, N.id = 50, missing = 0.05)
}
\arguments{
\item{N.files}{Positive integer. Number of files/courses/waves.}

\item{N.id}{Positive integer. Number of participants in each file/course/wave.}

\item{missing}{Numberic between \code{0} and \code{1} indicating the amount of missing data.}
}
\description{
This function may be used to simulate data in order to use \code{\link{aggregate_eval}} and \code{\link{plot_eval}} without real data.
}
\examples{
\dontrun{
# Simulate data for 250 participants, 50 in each of 5 courses
dat.x <- sim_eval(N.files = 5, N.id = 50, missing = 0)
}
}

